Added unit tests for HttpServer prefix handling and made HttpServer use IRouteHandler instead of RestRouteHandler directly.
Next step is to create the FileSystemRouteHandler so we can actually start serving files!
