+++ src/WebServer.UnitTests/WebServer.UnitTests.csproj
@@ -96,9 +96,10 @@
     <SDKReference Include="TestPlatform.Universal, Version=$(UnitTestPlatformVersion)" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="HttpServerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Rest\RestControllerRequestHandlerTest.cs" />
-    <Compile Include="RestWebServerTest.Happy.cs" />
+    <Compile Include="RestRouteHandlerTest.Happy.cs" />
     <Compile Include="TestHelpers\TestRestControllers.cs" />
     <Compile Include="UnitTestApp.xaml.cs">
       <DependentUpon>UnitTestApp.xaml</DependentUpon>
@@ -116,7 +117,7 @@
     <AppxManifest Include="Package.appxmanifest">
       <SubType>Designer</SubType>
     </AppxManifest>
-    <Compile Include="RestWebServerTest.Rainy.cs" />
+    <Compile Include="RestRouteHandlerTest.Rainy.cs" />
     <None Include="WebServer.UnitTests_TemporaryKey.pfx" />
   </ItemGroup>
   <ItemGroup>
