+++ src/WebServer/IRouteHandler.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using Devkoes.HttpMessage;
+
+namespace Devkoes.Restup.WebServer
+{
+    public interface IRouteHandler
+    {
+        Task<HttpServerResponse> HandleRequest(IHttpServerRequest request);
+    }
+}
