+++ src/WebServer/WebServer.csproj
@@ -115,6 +115,7 @@
     <Compile Include="Attributes\UriFormatAttribute.cs" />
     <Compile Include="Configuration.cs" />
     <Compile Include="EncodingCache.cs" />
+    <Compile Include="IRouteHandler.cs" />
     <Compile Include="Models\Schemas\RestServerRequest.cs" />
     <Compile Include="RestRoutehandler.cs" />
     <Compile Include="Rest\RestControllerMethodWithContentExecutor.cs" />
